+++ run.sh
@@ -3,4 +3,8 @@
 SRC_DIR="src"
 ENTRY_FILE="main.py"
 
+<<<<<<< HEAD
 python $SRC_DIR/$ENTRY_FILE
+=======
+python $SRC_DIR/$ENTRY_FILE
+>>>>>>> main