+++ src/copa_modules/__init__.py
@@ -1,1 +1,1 @@
-from .DataProcessor import DataProcessor
+from .data_processor import data_processor