+++ CHANGELOG.md
@@ -1,11 +1,21 @@
+<<<<<<< HEAD
 ## Unreleased
 
 ### Feat
 
+=======
+## 0.2.0 (2024-05-09)
+
+### Feat
+
+- **data_processor**: added fast load option
+- Save dataframe function added
+>>>>>>> main
 - **run.sh**: bash script to run the whole project
 
 ### Fix
 
+<<<<<<< HEAD
 - **pyproject.toml**: fixing the python version for macbooks arm chip
 - **pyproject.toml**: fixing the python version for macbooks arm chip
 - **poetry**: fixing the bug with cz bump
@@ -13,3 +23,14 @@
 ### Refactor
 
 - adding the changelog file
+=======
+- **gitignore**: added dataset file formats to git ignore
+- **pyproject.toml**: removed voex from dependencies
+- **pyproject.toml**: fixing the python version for macbooks arm chip
+
+### Refactor
+
+- removing parquet file
+
+## 0.0.1 (2024-05-08)
+>>>>>>> main